Added LionParameterizedTest, update LionTest.java
